requirements.txt
pyzmq
scikit-learn
numpy
